+++ tests/functional/s/super/super_init_not_called.py
@@ -3,8 +3,6 @@
 
 import ctypes
 
-from typing_extensions import Protocol as ExtensionProtocol
-
 
 class Foo(ctypes.BigEndianStructure):
     """This class should not emit a super-init-not-called warning.
@@ -15,24 +13,6 @@ class Foo(ctypes.BigEndianStructure):
 
     def __init__(self):
         ctypes.BigEndianStructure.__init__(self)
-
-
-class TestProto(ExtensionProtocol):
-    """A protocol without __init__ using Protocol from typing_extensions."""
-
-
-class TestParent(TestProto):
-    """An implementation."""
-
-    def __init__(self):
-        ...
-
-
-class TestChild(TestParent):
-    """An implementation which should call the init of TestParent."""
-
-    def __init__(self):  # [super-init-not-called]
-        ...
 
 
 class UninferableChild(UninferableParent):  # [undefined-variable]
