+++ tests/functional/s/super/super_init_not_called_extensions.rc
@@ -1,0 +1,2 @@
+[testoptions]
+max_pyver=3.9
