+++ tests/functional/s/super/super_init_not_called_extensions.txt
@@ -1,0 +1,1 @@
+super-init-not-called:21:4:22:11:TestChild.__init__:__init__ method from base class 'TestParent' is not called:INFERENCE
